test(j-s): move validation test to seperate file
